Fix asset tests in task_sdk that use db
Mocks are added to avoid real db access. The db entry point is made
lazy, and function expand_alias_to_assets moved to airflow.models and
tested there properly with db access.
