+++ task_sdk/tests/conftest.py
@@ -109,3 +109,22 @@ def captured_logs(request):
         yield cap.entries
     finally:
         structlog.configure(processors=cur_processors)
+
+
+@pytest.fixture(autouse=True, scope="session")
+def _disable_ol_plugin():
+    # The OpenLineage plugin imports setproctitle, and that now causes (C) level thread calls, which on Py
+    # 3.12+ issues a warning when os.fork happens. So for this plugin we disable it
+
+    # And we load plugins when setting the priorty_weight field
+    import airflow.plugins_manager
+
+    old = airflow.plugins_manager.plugins
+
+    assert old is None, "Plugins already loaded, too late to stop them being loaded!"
+
+    airflow.plugins_manager.plugins = []
+
+    yield
+
+    airflow.plugins_manager.plugins = None
